3.2.0

+++ lostandfind/src/main/java/com/example/lostandfind/Repository/SuggestRepository.java
@@ -3,5 +3,5 @@ package com.example.lostandfind.repository;
 import com.example.lostandfind.mysql.SuggestMysql;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SuggestRepository extends JpaRepository<SuggestMysql,Long> {
+public interface SuggestRepository extends JpaRepository<SuggestMysql,Integer> {
 }
